Split util headers
